@@ -31,22 +31,22 @@
     "release:docs": "hugo -s docs && gh-pages -d docs/public"
   },
   "dependencies": {
-    "space-router": "^0.8.4"
+    "space-router": "^0.8.5"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.0",
-    "@babel/preset-env": "^7.16.0",
-    "@babel/preset-react": "^7.16.0",
-    "@babel/register": "^7.16.0",
+    "@babel/cli": "^7.17.6",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/register": "^7.17.7",
     "ava": "^3.15.0",
     "gh-pages": "^3.2.3",
     "healthier": "^4.0.0",
-    "jsdom": "^18.1.0",
-    "np": "^7.5.0",
+    "jsdom": "^19.0.0",
+    "np": "^7.6.1",
     "nyc": "^15.1.0",
-    "prettier": "^2.4.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "prettier": "^2.6.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "healthier": {
     "ignore": [
